# sarc/docs/swagger.yaml
definitions:
...
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.Lecture:
    properties:
      classId:
        type: integer
      content:
        items:
          type: string
        type: array
      date:
        type: string
      lectureId:
        type: integer
      presence:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      roomId:
        type: integer
    type: object
  domain.Profile:
    properties:
      id:
        type: integer
      role:
        type: string
    type: object
...
  domain.User:
    properties:
      birthDate:
        type: string
      email:
        type: string
      id:
        type: integer
      nome:
        type: string
      profileId:
        type: integer
      sex:
        type: string
      telephone:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /lectures:
    get:
      description: Retrieves all lectures
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Lecture'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all lectures
      tags:
      - lectures
    post:
      consumes:
      - application/json
      description: Creates a new lecture in the system
      parameters:
      - description: Lecture data
        in: body
        name: lecture
        required: true
        schema:
          $ref: '#/definitions/domain.Lecture'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Lecture'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Create a new lecture
      tags:
      - lectures
  /lectures/{id}:
    delete:
      description: Deletes a lecture by its ID
      parameters:
      - description: Lecture ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Delete a lecture
      tags:
      - lectures
    get:
      description: Retrieves a lecture by its ID
      parameters:
      - description: Lecture ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Lecture'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Lecture not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get lecture by ID
      tags:
      - lectures
    put:
      consumes:
      - application/json
      description: Updates the lecture information for the given lecture ID
      parameters:
      - description: Lecture ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lecture data
        in: body
        name: lecture
        required: true
        schema:
          $ref: '#/definitions/domain.Lecture'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Lecture'
        "400":
          description: Invalid ID or bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update an existing lecture
      tags:
      - lectures
  /profiles:
    get:
      description: Retrieves all profiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Profile'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get all profiles
      tags:
      - profiles
    post:
      consumes:
      - application/json
      description: Creates a new profile in the system
      parameters:
      - description: Profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/domain.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Profile'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Create a new profile
      tags:
      - profiles
  /profiles/{id}:
    delete:
      description: Deletes a profile by its ID
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Delete a profile
      tags:
      - profiles
    get:
      description: Retrieves a profile by its ID
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Profile'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get profile by ID
      tags:
      - profiles
    put:
      consumes:
      - application/json
      description: Updates the profile information for the given profile ID
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/domain.Profile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Profile'
        "400":
          description: Invalid ID or bad request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update an existing profile
      tags:
      - profiles
...
swagger: "2.0"